src:nnet/nnet-loss.h prevent overflow when large data set is applied
git-svn-id: https://svn.code.sf.net/p/kaldi/code/trunk@4726 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/nnet/nnet-loss.h
@@ -94,7 +94,7 @@ class Mse {
   std::string Report();
 
  private:
-  int32 frames_;
+  int64 frames_;
   double loss_;
   
   int32 frames_progress_;
